cleaned up after making Point

+++ Pong/utils.hpp
@@ -1,7 +1,8 @@
 #pragma once
 
 #include "olcPixelGameEngine.h"
+
 struct Point {
-    int x;
-    int y;
+    float x;
+    float y;
 };
